Removed superfluous total text shift from child classes
