+++ lib/message.rb
@@ -50,19 +50,19 @@ class Message
   #   end
   # end
 
-  # def shift_text_array(key)
-  #   shift_text_array_specific(key, @a_array, key.a_shift, @a_shifted)
-  #   shift_text_array_specific(key, @b_array, key.b_shift, @b_shifted)
-  #   shift_text_array_specific(key, @c_array, key.c_shift, @c_shifted)
-  #   shift_text_array_specific(key, @d_array, key.d_shift, @d_shifted)
-  # end
-
-  def reverse_shift_text_array(key)
-    reverse_shift_text_array_specific(key, @a_array, key.a_shift, @a_shifted)
-    reverse_shift_text_array_specific(key, @b_array, key.b_shift, @b_shifted)
-    reverse_shift_text_array_specific(key, @c_array, key.c_shift, @c_shifted)
-    reverse_shift_text_array_specific(key, @d_array, key.d_shift, @d_shifted)
+  def shift_text_array(key)
+    shift_text_array_specific(key, @a_array, key.a_shift, @a_shifted)
+    shift_text_array_specific(key, @b_array, key.b_shift, @b_shifted)
+    shift_text_array_specific(key, @c_array, key.c_shift, @c_shifted)
+    shift_text_array_specific(key, @d_array, key.d_shift, @d_shifted)
   end
+
+  # def reverse_shift_text_array(key)
+  #   reverse_shift_text_array_specific(key, @a_array, key.a_shift, @a_shifted)
+  #   reverse_shift_text_array_specific(key, @b_array, key.b_shift, @b_shifted)
+  #   reverse_shift_text_array_specific(key, @c_array, key.c_shift, @c_shifted)
+  #   reverse_shift_text_array_specific(key, @d_array, key.d_shift, @d_shifted)
+  # end
 
   def reform_shifted_text(key)
     @shifted_text = make_shifted_array(key).join
