+++ test/decrypted_message_test.rb
@@ -20,15 +20,4 @@ class DecryptedMessageTest < Minitest::Test
 
     assert_equal ["e"," ","l"], @reverse_message.b_shifted
   end
-
-  def test_it_can_reverse_shift_all_text_arrays_simultaneously
-    @reverse_message.text_to_array
-    @key.letter_shift
-    @reverse_message.reverse_shift_text_array(@key)
-
-    assert_equal ["h","o","r"], @reverse_message.a_shifted
-    assert_equal ["e"," ","l"], @reverse_message.b_shifted
-    assert_equal ["l","w","d"], @reverse_message.c_shifted
-    assert_equal ["l","o"], @reverse_message.d_shifted
-  end
 end
