+++ test/encrypted_message_test.rb
@@ -20,16 +20,4 @@ class EncryptedMessageTest < Minitest::Test
 
     assert_equal ["j","e","q"], @message.b_shifted
   end
-
-  def test_it_can_shift_all_text_arrays_simultaneously
-    @message.text_to_array
-    @key.letter_shift
-    @message.shift_text_array(@key)
-    # require "pry"; binding.pry
-
-    assert_equal ["a", "h", "k"], @message.a_shifted
-    assert_equal ["j","e","q"], @message.b_shifted
-    assert_equal ["w", "g", "o"], @message.c_shifted
-    assert_equal ["g", "j"], @message.d_shifted
-  end
 end
